changing status hardcoded

+++ src/main/java/io/lab/hotel/iPresenter.java
@@ -10,7 +10,7 @@ public interface iPresenter {
 
 	boolean acceptHotel(int hotelID);
 
-	boolean changeStatus(float reservationID);
+	boolean changeStatus(int reservationID);
 
 	/**
 	 *
